third_party/blink/web_tests/external/wpt/css/css-scoping/font-face-006-expected.txt

This is a testharness.js-based test.
[FAIL] @font-face from document applies to to :host::before/::after.
  assert_greater_than: expected a number greater than 160 but got 84
Harness: the test ran to completion.
